--- crawler/input/InputURLs.txt
@@ -1,2 +1,0 @@
-#insert urls here
-#file can be changed to whatever filetype the assigned programmer desires (csv, yaml, txt)
