--- crawler/input/clients.yaml
@@ -1,10 +1,0 @@
-#possible user clients: iphone, android, chrome_windows, chrome_macintosh,
-#firefox_windows, firefox_macintosh, safari, linux
-safari: safari
-iphone: iphone
-android: android
-chrome_windows: chrome_windows
-chrome_macintosh: chrome_macintosh
-firefox_windows: firefox_windows
-firefox_macintosh: firefox_macintosh
-linux: linux
