+++ crawler/tests/test_headers.py
@@ -1,8 +1,7 @@
 from crawler.header_creater.create_header import generate_header
-from crawler.header_creater.user_agent_creator import *
+from crawler.log import logging_config
 
-
-test_map = {"android"}
+test_map = "android"
 
 test_header = generate_header(test_map)
-print(test_header)
+logging_config.warning(test_header)
